+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 ## [@project-version@] - @project-date-iso@
 ### New Features
 - Datatext Drawer - Click one of the four corners of the screen to open the datatext drawer. Drag and drop them anywhere on-screen. Drag them back into the drawer to disable. Drag the right hand corner to resize them. Will continue to expand on these and you're welcome to suggest more/improve whats been built.
+
 - Meta Talents - Added a new QoL feature to help you keep track of the most popular talents for each spec. Find it in the Quality of Life tab in `/orbit plugins`.
     - Select bosses or dungeons to view the most popular talents for that specific encounter, data is fetched and averaged from Warcraft Logs top 100 parses. 
     - Directly Apply the meta talents to your talent tree with a single click.
 
 ### Bugfixes
 - Static Cooldown Timer Texts
+- Player Buff Item Enhancements now draw a pixel border for item enhancements (weapon oils, etc)
+- Player Buffs/Debuffs swipe now start at low alpha and fill as their duration expires
+- Player Buffs/Debuffs now pulse in and out when expiring
 
 ## [1.0.0] - 2026-03-10
 ### Added